structure functions in different files for middleware and schema
