+++ apps/discovery-platform/src/entry-server/routers/v1/index.ts
@@ -1,6 +1,5 @@
-import express, { NextFunction, Request, Response } from "express";
+import express, { Request, Response } from "express";
 import {
-  ParamSchema,
   body,
   checkSchema,
   param,
@@ -15,7 +14,7 @@ import {
 } from "../../../client";
 import { DBInstance } from "../../../db";
 import { FundingServiceApi } from "../../../funding-service-api";
-import { createQuota, getSumOfQuotasPaidByClient } from "../../../quota";
+import { createQuota } from "../../../quota";
 import {
   createRegisteredNode,
   getEligibleNode,
@@ -24,142 +23,18 @@ import {
 } from "../../../registered-node";
 import { ClientDB, RegisteredNode } from "../../../types";
 import { createLogger, isListSafe } from "../../../utils";
-import { Histogram } from "prom-client";
-import memoryCache from "memory-cache";
 import { errors } from "pg-promise";
 import { MetricManager } from "@rpch/common/build/internal/metric-manager";
+import * as constants from "../../../constants";
+import { getNodeSchema, registerNodeSchema } from "./schema";
+import {
+  doesClientHaveQuota,
+  getCache,
+  metricMiddleware,
+  setCache,
+} from "./middleware";
 
 const log = createLogger(["entry-server", "router", "v1"]);
-
-// payment mode when quota is paid by trial
-const TRIAL_PAYMENT_MODE = "trial";
-
-// client id that will pay for quotas in trial mode
-const TRIAL_CLIENT_ID = "trial";
-
-// Sanitization and Validation
-const registerNodeSchema: Record<keyof RegisteredNode, ParamSchema> = {
-  peerId: {
-    in: "body",
-    exists: {
-      errorMessage: "Expected peerId to be in the body",
-      bail: true,
-    },
-    isString: true,
-  },
-  chainId: {
-    in: "body",
-    exists: {
-      errorMessage: "Expected chainId to be in the body",
-      bail: true,
-    },
-    isNumeric: true,
-    toInt: true,
-  },
-  exitNodePubKey: {
-    in: "body",
-    exists: {
-      errorMessage: "Expected exitNodePubKey to be in the body",
-      bail: true,
-    },
-    isString: true,
-  },
-  hasExitNode: {
-    in: "body",
-    exists: {
-      errorMessage: "Expected hasExitNode to be in the body",
-      bail: true,
-    },
-    isBoolean: true,
-    toBoolean: true,
-  },
-  hoprdApiEndpoint: {
-    in: "body",
-    exists: {
-      errorMessage: "Expected hoprdApiEndpoint to be in the body",
-      bail: true,
-    },
-    isString: true,
-  },
-  hoprdApiToken: {
-    in: "body",
-    exists: {
-      errorMessage: "Expected hoprdApiToken to be in the body",
-      bail: true,
-    },
-    isString: true,
-  },
-  nativeAddress: {
-    in: "body",
-    exists: {
-      errorMessage: "Expected nativeAddress to be in the body",
-      bail: true,
-    },
-    isString: true,
-  },
-};
-
-const getNodeSchema: Record<
-  keyof { excludeList?: string; hasExitNode?: string },
-  ParamSchema
-> = {
-  excludeList: {
-    optional: true,
-    in: "query",
-    custom: {
-      options: (value) => {
-        return isListSafe(value);
-      },
-    },
-  },
-  hasExitNode: {
-    optional: true,
-    in: "query",
-    isBoolean: true,
-  },
-};
-
-export const getCache = () => {
-  return (req: Request, res: Response<any, any>, next: NextFunction) => {
-    let key = req.originalUrl || req.url;
-    let cachedBody = memoryCache.get(key);
-    if (cachedBody) {
-      log.verbose("Returning cached value for endpoint: ", key);
-      return res.json(cachedBody);
-    }
-    next();
-  };
-};
-
-export const setCache = (key: string, duration: number, body: Object) => {
-  memoryCache.put(key, body, duration);
-};
-
-// middleware used to check if client sent in req has enough quota
-export const doesClientHaveQuota = async (
-  db: DBInstance,
-  client: string,
-  baseQuota: bigint
-) => {
-  const sumOfClientsQuota = await getSumOfQuotasPaidByClient(db, client);
-  return sumOfClientsQuota >= baseQuota;
-};
-
-// middleware that will track duration of request
-export const metricMiddleware =
-  (histogramMetric: Histogram<string>) =>
-  (req: Request, res: Response, next: NextFunction) => {
-    const start = process.hrtime();
-    res.on("finish", () => {
-      const end = process.hrtime(start);
-      const durationSeconds = end[0] + end[1] / 1e9;
-      const statusCode = res.statusCode.toString();
-      const method = req.method;
-      const path = req.path;
-      histogramMetric.labels(method, path, statusCode).observe(durationSeconds);
-    });
-    next();
-  };
 
 // Express Router
 export const v1Router = (ops: {
@@ -320,7 +195,7 @@ export const v1Router = (ops: {
     }
   );
 
-  // DISCLAIMER: can be exploited to allow client to use infinite funds
+  // TODO: can be exploited to allow client to use infinite funds
   router.post(
     "/client/quota",
     metricMiddleware(requestDurationHistogram),
@@ -352,7 +227,7 @@ export const v1Router = (ops: {
 
         if (!dbClient) throw Error("Could not create Client");
 
-        if (dbClient.payment === TRIAL_PAYMENT_MODE) {
+        if (dbClient.payment === constants.TRIAL_PAYMENT_MODE) {
           // update client to premium of it was previously trial
           await updateClient(ops.db, { ...dbClient, payment: "premium" });
         }
@@ -445,9 +320,12 @@ export const v1Router = (ops: {
           });
         }
 
-        const clientIsTrialMode = dbClient?.payment === TRIAL_PAYMENT_MODE;
+        const clientIsTrialMode =
+          dbClient?.payment === constants.TRIAL_PAYMENT_MODE;
         // set who is going to pay for quota
-        const paidById = clientIsTrialMode ? TRIAL_CLIENT_ID : dbClient?.id;
+        const paidById = clientIsTrialMode
+          ? constants.TRIAL_CLIENT_ID
+          : dbClient?.id;
 
         // check if client has enough quota
         const doesClientHaveQuotaResponse = await doesClientHaveQuota(
